+++ src/simulation/natto_simple_simulation/src/swerve_simulator.cpp
@@ -75,7 +75,7 @@ void swerve_simulator::timer_callback () {
     natto_msgs::msg::Swerve command_sum;
     command_sum.wheel_angle.resize (num_wheels_, 0.0);
     command_sum.wheel_speed.resize (num_wheels_, 0.0);
-    for (int i = 0; i < received_commands_.size (); i++) {
+    for (size_t i = 0; i < received_commands_.size (); i++) {
         for (int j = 0; j < num_wheels_; j++) {
             if (received_commands_[i].wheel_speed.size () != num_wheels_ || received_commands_[i].wheel_angle.size () != num_wheels_) {
                 RCLCPP_FATAL (this->get_logger (), "Received command size does not match number of wheels.");
